requirements.txt
accelerate==1.5.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.0
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
applang==0.2.2
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.4
blake3==1.0.4
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorful==0.5.6
compressed-tensors==0.9.1
contourpy==1.3.1
cryptography==44.0.2
cuda-bindings==12.8.0
cuda-python==12.8.0
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.4.1
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
diffusers==0.33.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
easydict==1.13
einops==0.8.1
executing==2.2.0
faiss-cpu==1.7.4
faiss-gpu-cu12==1.11.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.18.0
flash_attn==2.7.4.post1
flashinfer-python==0.2.3+cu124torch2.5
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gdown==5.2.0
gguf==0.10.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.2
hf_transfer==0.1.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
interegular==0.3.3
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonargparse==4.37.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lark==1.2.2
lean-dojo==2.2.0
libcst==1.7.0
litellm==1.63.12
llguidance==0.7.4
lm-format-enforcer==0.10.11
loguru==0.7.3
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mistune==3.1.3
modelscope==1.24.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
mypackage==0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.67.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
outcome==1.3.0.post0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
roman==5.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.5
sgl-kernel==0.0.5
sglang==0.4.4.post1
shtab==1.7.2
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
termcolor2==0.0.3
threadpoolctl==3.6.0
tiktoken==0.9.0
time-machine==2.16.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1
torchao==0.9.0
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.1.0
trl==0.15.2
typeguard==4.4.2
types-psutil==7.0.0.20250218
types-toml==0.10.8.20240310
typing_extensions==4.13.2
tyro==0.9.21
tzdata==2025.1
unsloth==2025.5.7
unsloth_zoo==2025.5.8
urllib3==2.3.0
uv==0.6.8
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
vllm==0.7.2
wandb==0.19.11
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.15
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
zulip==0.9.0
